Python/audio files/requirements.txt
deepspeech-gpu==0.9.3
gputil==1.4.0
pytest==7.2.0